refactor: upgrade dev deps
