+++ package.json
@@ -206,7 +206,7 @@
     "react-icons": "4.8.0",
     "react-lottie": "1.2.3",
     "react-redux": "8.0.5",
-    "react-router-dom": "6.11.1",
+    "react-router-dom": "6.11.2",
     "react-virtuoso": "4.0.8",
     "redux-persist": "6.0.0",
     "rxjs": "7.8.1",
@@ -229,7 +229,7 @@
     "@emotion/babel-preset-css-prop": "11.11.0",
     "@emotion/cache": "11.11.0",
     "@ls-lint/ls-lint": "1.11.2",
-    "@playwright/test": "1.30.0",
+    "@playwright/test": "1.34.0",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@redux-devtools/cli": "2.0.0",
     "@redux-devtools/remote": "0.8.0",
@@ -285,7 +285,7 @@
     "cross-env": "7.0.3",
     "crypto-browserify": "3.12.0",
     "deepmerge": "4.3.1",
-    "dependency-cruiser": "13.0.1",
+    "dependency-cruiser": "13.0.2",
     "dotenv-webpack": "8.0.1",
     "esbuild": "0.17.19",
     "esbuild-loader": "3.0.1",
@@ -300,7 +300,6 @@
     "jest-circus": "29.3.1",
     "jest-dev-server": "6.2.0",
     "jsdom": "21.1.0",
-    "playwright": "1.30.0",
     "prettier": "2.8.8",
     "process": "0.11.10",
     "progress-bar-webpack-plugin": "2.1.0",
@@ -319,7 +318,7 @@
     "vm-browserify": "1.1.2",
     "web-ext": "7.4.0",
     "web-ext-submit": "7.4.0",
-    "webpack": "5.82.1",
+    "webpack": "5.83.1",
     "webpack-bundle-analyzer": "4.8.0",
     "webpack-cli": "5.1.1",
     "webpack-dev-server": "4.15.0",
