+++ tests/mocks/mock-apis.ts
@@ -1,5 +1,5 @@
+import { Page } from '@playwright/test';
 import { json } from '@tests/utils';
-import { Page } from 'playwright';
 
 export const setupMockApis = async (page: Page) => {
   await page.route(/chrome-extension/, route => route.continue());
